# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TuMobilya-API",
        "description": "TuMobilya-API için kimlik doğrulama ve mobilya yönetimi API'si",
        "contact": {
            "email": "info@tumobilya.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Yerel geliştirme sunucusu"
        },
        {
            "url": "http://tumobilya-api.test"
        }
    ],
    "paths": {
        "/api/announcements": {
            "get": {
                "tags": [
                    "Duyurular"
                ],
                "summary": "Tüm duyuruları listele",
                "description": "Kayıtlı tüm duyuruları döndürür.",
                "operationId": "71c08f9cc8bee71c81626251da258590",
                "responses": {
                    "200": {
                        "description": "Duyurular başarıyla listelendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Yeni Mağaza Açılışı"
                                            },
                                            "content": {
                                                "type": "string",
                                                "example": "İstanbul'da yeni mağazamız açıldı!"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "announcement1.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kimlik doğrulanmadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Duyurular"
                ],
                "summary": "Yeni duyuru ekle",
                "description": "Yeni bir duyuru oluşturur.",
                "operationId": "dc2f49bef808056e16e168f71b8830ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Yıl Sonu İndirimi"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Tüm ürünlerde %30 indirim fırsatı!"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "announcement2.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Duyuru başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Yıl Sonu İndirimi"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Tüm ürünlerde %30 indirim fırsatı!"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "announcement2.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Duyuru oluşturuldu"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The title field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/announcements/{id}": {
            "get": {
                "tags": [
                    "Duyurular"
                ],
                "summary": "Tek bir duyuruyu getir",
                "description": "Belirtilen ID'ye sahip duyuruyu döndürür.",
                "operationId": "2938bef7a01a39cf89e059206f5ae0d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duyuru bulundu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Yeni Mağaza Açılışı"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "İstanbul'da yeni mağazamız açıldı!"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "announcement1.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Duyuru bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kaynak bulunamadı"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Duyurular"
                ],
                "summary": "Duyuru güncelle",
                "description": "Belirtilen ID'ye sahip duyuruyu günceller.",
                "operationId": "a8df15398083d218b208ceabff089eb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Yıl Sonu İndirimi (Güncellendi)"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Tüm ürünlerde %35 indirim fırsatı!"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "announcement2-updated.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Duyuru güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Yıl Sonu İndirimi (Güncellendi)"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Tüm ürünlerde %35 indirim fırsatı!"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "announcement2-updated.jpg"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Duyuru güncellendi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Duyuru bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kaynak bulunamadı"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Duyurular"
                ],
                "summary": "Duyuru sil",
                "description": "Belirtilen ID'ye sahip duyuruyu siler.",
                "operationId": "6e6001a5b104477475d1a70294d42ba1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Duyuru silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Duyuru silindi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Duyuru bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kaynak bulunamadı"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}